src/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
pandas==2.2.1
fuzzywuzzy==0.18.0
SQLAlchemy==2.0.27
numpy==1.26.4
Levenshtein==0.25.0
jinja2==3.1.3
